新增骑手端, 实现管理-骑手-用户的闭环！！！
 1. 高并发抢单与数据一致性
- 并发流量防线：引入 Redisson 分布式锁 (RLock) 实现抢单业务的细粒度互斥，配合 tryLock 快速失败机制，有效拦截瞬时高并发流量，防止数据库击穿。
- 数据最终一致性：在数据库层采用 乐观锁 机制，作为数据安全的最后一道防线，彻底杜绝“一单多抢”现象。
- 独立鉴权体系：构建基于 JWT + Interceptor 的骑手端独立认证模块，实现与管理端/用户端的权限物理隔离。

 2. 实时交互与消息触达
- 毫秒级消息推送：集成 WebSocket 全双工通信服务，打通“商家接单 -> 服务端事件 -> 骑手端广播”的实时链路，替代传统轮询方案，显著降低服务器负载并提升接单响应速度。
- 抢单大厅：重构订单查询逻辑，实现基于状态机的订单池流转。

 3. 配送安全与业务闭环
- 取餐码机制 (Pickup Code)：设计包含“下单生成 -> 抢单可见 -> 校验核销”的全生命周期取餐码逻辑，解决校园/外卖场景下的错拿、漏拿问题。
- 状态机流转控制：完善订单状态流转的严密性校验（如：仅限本人操作、状态前置检查），防止越权操作与状态跳变。
- 作业流闭环：实现“我的配送列表”、“确认送达”、“今日数据统计”等核心履约功能，完成外卖业务的最后一块拼图。

+++ campus-server/src/main/java/com/campus/mapper/PermissionMapper.java
@@ -12,6 +12,6 @@ public interface PermissionMapper {
     /*
      * 根据roleId查询权限信息
      * */
-    List<String> selectPermissionByRoleId(Integer roleId);
+    List<String> selectPermissionByRoleId(Long roleId);
 
 }
