+++ campus-server/src/main/java/com/campus/mapper/RoleMapper.java
@@ -13,5 +13,5 @@ public interface RoleMapper {
      * 查询角色
      * */
     @Select("select code from campus_delivery.sys_role where id = #{roleId}")
-    List<String> selectCodeByRoleId(Integer roleId);
+    List<String> selectCodeByRoleId(Long roleId);
 }
